update item types
this was development hell holy shit my brain is smoking

so main gist now is to either enable or disable modded enchantments canApply or canApplyAtEnchantingTable behavior, which might or might not come back on track to Enchantment.canApply/AtEnch. players can modify by enchantment or in general.

defaults try to be informative and not forcing modified behavior. no guarantees though that it doesnt modify some weird modded behavior anyway